+++ srcs/ft_calc_mb.c
@@ -6,18 +6,18 @@
 /*   By: cledant <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/19 15:43:44 by cledant           #+#    #+#             */
-/*   Updated: 2016/07/28 21:28:57 by cledant          ###   ########.fr       */
+/*   Updated: 2016/08/03 21:00:36 by cledant          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fract_ol.h"
 
-int		ft_calc_mb(float *init_x, float *init_y, float x, float y)
+int		ft_calc_mb(double *init_x, double *init_y, double x, double y)
 {
-	float	tmp_x;
-	float	tmp_y;
-	float	sqrt_tmp_x;
-	float	sqrt_tmp_y;
+	double	tmp_x;
+	double	tmp_y;
+	double	sqrt_tmp_x;
+	double	sqrt_tmp_y;
 
 	tmp_x = *init_x;
 	tmp_y = *init_y;
