+++ srcs/ft_init_cuda.c
@@ -6,7 +6,7 @@
 /*   By: cledant <cledant@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/08/02 16:37:10 by cledant           #+#    #+#             */
-/*   Updated: 2016/08/03 18:54:20 by cledant          ###   ########.fr       */
+/*   Updated: 2016/08/03 21:01:30 by cledant          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,19 +39,19 @@ int				ft_cuda_init(t_mlx *e)
 	bigger = (e->win_x_size > e->win_y_size) ? e->win_x_size : e->win_y_size;
 	e->nb_block = (bigger % M_THREAD == 0) ? bigger / M_THREAD :
 		bigger / M_THREAD + 1;
-	if (cudaMalloc((void **)(&e->d_x_min), sizeof(float)) != cudaSuccess)
+	if (cudaMalloc((void **)(&e->d_x_min), sizeof(double)) != cudaSuccess)
 		return (0);
-	if (cudaMalloc((void **)(&e->d_y_max), sizeof(float)) != cudaSuccess)
+	if (cudaMalloc((void **)(&e->d_y_max), sizeof(double)) != cudaSuccess)
 	{
 		ft_clean(e, 1);
 		return (0);
 	}
-	if (cudaMalloc((void **)(&e->d_x_pitch), sizeof(float)) != cudaSuccess)
+	if (cudaMalloc((void **)(&e->d_x_pitch), sizeof(double)) != cudaSuccess)
 	{
 		ft_clean(e, 2);
 		return (0);
 	}
-	if (cudaMalloc((void **)(&e->d_y_pitch), sizeof(float)) != cudaSuccess)
+	if (cudaMalloc((void **)(&e->d_y_pitch), sizeof(double)) != cudaSuccess)
 	{
 		ft_clean(e, 3);
 		return (0);
